:truck: Rename: 메인 홈 페이지 이름 변경
